fixed lexing of numbers
The number scanner (ifstream_lexer::scan_number()) was eating
one extra character from the input when leaving its feed loop.
The condition to stop feeding input into the number token is to
find a non-numeric char (or exahusting input). When a non-numeric
char is found, the loop finishes but this char has been already consumed
by the lexer. To continue lexing the input we have to put that char back
into the input