Ignore git tags in llvm-config version numbers
Some llvm precompiled releases, such as 5.0.0 have a git tag as part of
the version number given by their llvm-config, but the rest of the setup
(for example, the system includes directory) uses the semantic
versioning tag ("5.0.0") only.

This commit adds a regex filter to ignore any tag appended to the llvm
version number, if any.