Use cppast specific cmake function for updating git submodules
The type_safe fork also uses a `configure_submodule()`
cmake function. This causes the cppast function to be hidden by the
type_safe one after type_safe is configured and added to the project.
When this happens, initialization of the following cppast submodule
fails since these no longer use the cppast function but the type_safe
one, which works from the type_safe repo as working directory.